requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==5.2.0
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2.8
django-cors-headers==3.13.0
django-phonenumber-field==6.3.0
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
joblib==1.2.0
nltk==3.7
oauthlib==3.2.1
phonenumbers==8.12.54
Pillow==9.2.0
protobuf==4.21.7
psycopg2-binary==2.9.3
py-w3c==0.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tinycss2==1.1.1
tqdm==4.64.1
ultimate-sitemap-parser==0.5
uritemplate==4.1.1
urllib3==1.26.10
webencodings==0.5.1
aiohttp==3.8.3
mimetype==0.1.5
xmltodict==0.13.0
django-widget-tweaks==1.4.12
django-bootstrap-v5==1.0.11
django-ckeditor==6.5.1
transformers==4.28.1
torch==2.0.0+cpu
torchaudio==2.0.1+cpu
torchvision==0.15.1+cpu
tqdm==4.64.1